+++ src/routes/examples/events/blur.svelte
@@ -1,0 +1,18 @@
+<script>
+    import Select from '$lib/Select.svelte';
+
+    let items = [
+        { value: 'one', label: 'One' },
+        { value: 'two', label: 'Two' },
+        { value: 'three', label: 'Three' },
+    ];
+
+    function handle(e) {
+        console.log(e.detail);
+    }
+</script>
+
+<Select {items} on:blur={handle} />
+
+
+<button>Next</button>