+++ src/routes/examples/props/placeholder-always-show.svelte
@@ -1,0 +1,11 @@
+<script>
+    import Select from '$lib/Select.svelte';
+
+    let items = [
+        { value: 'one', label: 'One' },
+        { value: 'two', label: 'Two' },
+        { value: 'three', label: 'Three' },
+    ];
+</script>
+
+<Select {items} placeholderAlwaysShow multiple />