+++ src/routes/examples/slots/chevron-icon.svelte
@@ -1,0 +1,14 @@
+<script>
+    import Select from '$lib/Select.svelte';
+
+    let items = [
+        { value: 'one', label: 'One' },
+        { value: 'two', label: 'Two' },
+        { value: 'three', label: 'Three' },
+    ];
+</script>
+
+
+<Select {items} showChevron>
+    <div slot="chevron-icon">⬇️</div>
+</Select>