+++ src/routes/index.svelte
@@ -2,7 +2,63 @@
     import Select from '$lib/Select.svelte';
 </script>
 
-<Select />
+<h1>Examples</h1>
+
+<h2>Props</h2>
+<ul>
+    <li><a href="/examples/props/value">value</a></li>
+    <li><a href="/examples/props/label">label</a></li>
+    <li><a href="/examples/props/just-value">justValue</a></li>
+    <li><a href="/examples/props/items">items</a></li>
+    <li><a href="/examples/props/item-id">itemId</a></li>
+    <li><a href="/examples/props/id">id</a></li>
+    <li><a href="/examples/props/filter-text">filterText</a></li>
+    <li><a href="/examples/props/placeholder">placeholder</a></li>
+    <li><a href="/examples/props/hide-empty-state">hideEmptyState</a></li>
+    <li><a href="/examples/props/list-open">listOpen</a></li>
+    <li><a href="/examples/props/class">class</a></li>
+    <li><a href="/examples/props/container-styles">containerStyles</a></li>
+    <li><a href="/examples/props/clearable">clearable</a></li>
+    <li><a href="/examples/props/disabled">disabled</a></li>
+    <li><a href="/examples/props/multiple">multiple</a></li>
+    <li><a href="/examples/props/searchable">searchable</a></li>
+    <li><a href="/examples/props/group-header-selectable">groupHeaderSelectable</a></li>
+    <li><a href="/examples/props/focused">focused</a></li>
+    <li><a href="/examples/props/list-placement">listPlacement</a></li>
+    <li><a href="/examples/props/list-auto-width">listAutoWidth</a></li>
+    <li><a href="/examples/props/show-chevron">showChevron</a></li>
+    <li><a href="/examples/props/input-attributes">inputAttributes</a></li>
+    <li><a href="/examples/props/placeholder-always-show">placeholderAlwaysShow</a></li>
+    <li><a href="/examples/props/loading">loading</a></li>
+    <li><a href="/examples/props/list-offset">listOffset</a></li>
+    <li><a href="/examples/props/debounce-wait">debounceWait</a></li>
+    <li><a href="/examples/props/append-list-target">appendListTarget</a></li>
+</ul>
+
+<h2>Slots</h2>
+<ul>
+    <li><a href="/examples/slots/prepend">prepend</a></li>
+    <li><a href="/examples/slots/selection">selection</a></li>
+    <li><a href="/examples/slots/clear-icon">clear-icon</a></li>
+    <li><a href="/examples/slots/multi-clear-icon">multi-clear-icon</a></li>
+    <li><a href="/examples/slots/loading-icon">loading-icon</a></li>
+    <li><a href="/examples/slots/chevron-icon">chevron-icon</a></li>
+    <li><a href="/examples/slots/list">list</a></li>
+    <li><a href="/examples/slots/item">item</a></li>
+    <li><a href="/examples/slots/empty">empty</a></li>
+</ul>
+
+<h2>Events</h2>
+<ul>
+    <li><a href="/examples/events/select">select</a></li>
+    <li><a href="/examples/events/change">change</a></li>
+    <li><a href="/examples/events/focus">focus</a></li>
+    <li><a href="/examples/events/blur">blur</a></li>
+    <li><a href="/examples/events/clear">clear</a></li>
+    <li><a href="/examples/events/loaded">loaded</a></li>
+    <li><a href="/examples/events/item-created">itemCreated</a></li>
+    <li><a href="/examples/events/error">error</a></li>
+</ul>
 
 <style>
 </style>